reverse DB suffix table names, move OutputFile to output_file.py
