+++ src/ophys_etl/workflows/pipeline_modules/roi_classification/generate_correlation_projection.py
@@ -11,7 +11,9 @@ from ophys_etl.workflows.app_config.app_config import app_config
 
 from ophys_etl.workflows.ophys_experiment import OphysExperiment
 
-from ophys_etl.workflows.pipeline_module import PipelineModule, OutputFile
+from ophys_etl.workflows.output_file import OutputFile
+
+from ophys_etl.workflows.pipeline_module import PipelineModule
 
 
 class GenerateCorrelationProjectionModule(PipelineModule):
