+++ tests/workflows/pipeline_modules/test_segmentation.py
@@ -15,8 +15,8 @@ setup_app_config(
     'di_model.h5'
 )
 
-from ophys_etl.workflows.db.schemas import OphysROI, OphysROIMaskValueDB  # noqa E402
-from ophys_etl.workflows.pipeline_module import OutputFile  # noqa E402
+from ophys_etl.workflows.db.schemas import OphysROI, OphysROIMaskValue  # noqa E402
+from ophys_etl.workflows.output_file import OutputFile  # noqa E402
 
 from ophys_etl.workflows.db.db_utils import save_job_run_to_db  # noqa E402
 from sqlmodel import create_engine, Session, select # noqa E402
@@ -53,7 +53,7 @@ class TestSegmentation(MockSQLiteDB):
             )
         with Session(self._engine) as session:
             rois = session.exec(select(OphysROI)).all()
-            masks = session.exec(select(OphysROIMaskValueDB)).all()
+            masks = session.exec(select(OphysROIMaskValue)).all()
 
         assert len(rois) == 2
         assert rois[0].x == 19
