+++ tests/workflows/test_pipeline_module.py
@@ -22,8 +22,9 @@ from typing import List, Dict   # noqa #402
 
 from ophys_etl.workflows.ophys_experiment import OphysExperiment, \
     OphysSession, Specimen  # noqa #402
+from ophys_etl.workflows.output_file import OutputFile
 from ophys_etl.workflows.pipeline_module import PipelineModule, \
-    ModuleOutputFileExistsError, OutputFile  # noqa #402
+    ModuleOutputFileExistsError  # noqa #402
 
 
 class _DummyMod(PipelineModule):
