+++ tests/workflows/test_workflow_step_runs.py
@@ -17,14 +17,14 @@ setup_app_config(
 
 from ophys_etl.workflows.well_known_file_types import WellKnownFileType # noqa E402
 
-from ophys_etl.workflows.pipeline_module import OutputFile  # noqa E402
+from ophys_etl.workflows.output_file import OutputFile  # noqa E402
 from ophys_etl.workflows.workflow_step_runs import get_latest_run, \
     get_well_known_file_for_latest_run  # noqa E402
 from ophys_etl.workflows.workflow_steps import WorkflowStep # noqa E402
 
 from ophys_etl.workflows.db.db_utils import save_job_run_to_db  # noqa E402
 
-from ophys_etl.workflows.db.initialize_db import IntializeDBRunner  # noqa E402
+from ophys_etl.workflows.db.initialize_db import InitializeDBRunner  # noqa E402
 from sqlmodel import Session    # noqa E402
 
 from ophys_etl.workflows.workflow_names import WorkflowName # noqa E402
@@ -38,7 +38,7 @@ class TestWorkflowStepRuns:
         os.makedirs(cls._db_path.parent, exist_ok=True)
 
         db_url = f'sqlite:///{cls._db_path}'
-        cls._engine = IntializeDBRunner(
+        cls._engine = InitializeDBRunner(
             input_data={
                 'db_url': db_url
             },
